# Web/Clase3TT/Practica/docs/swagger.yaml
definitions:
  handler.request:
    properties:
      activo:
        type: boolean
      altura:
        type: number
      apellido:
        type: string
      edad:
        type: integer
      email:
        type: string
      fechaDeCreacion:
        type: string
      nombre:
        type: string
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
    type: object
info:
  contact:
    name: API Support
    url: https://developers.mercadolibre.com.ar/support
  description: Esto es una descripcion.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: Esto es un titulo
  version: "1.0"
paths:
  /products:
    get:
      consumes:
      - application/json
      description: get products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: List products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: store products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to store
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Store products
      tags:
      - Products
  /usuarios:
    patch:
      consumes:
      - application/json
      description: modificar usuarios
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Usuario a modificar
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Modificar usuarios
      tags:
      - Usuarios
swagger: "2.0"